fix Settings test and refactor folder structures
