+++ apps/settings/js/panels/phone_lock_dialog/phone_lock_dialog.js
